RF(BM): shorter names, separate files
was hard to navigate benchmarks results.  now master branch should be first
and then benchmark class/module names short
